Factor out FileStore.
